+++ src/components/common/header/nav.tsx
@@ -1,8 +1,12 @@
 import styles from '@/styles/components/common/header/header.module.scss';
 import { PageLink } from '@/src/components/common/links/page-link';
-import { User } from '@/src/components/common/user';
+import AccountDropdown from '@/src/components/common/account-dropdown';
+import { useUser } from '@/src/hooks/use-user';
+import { Button } from '@/src/components/common/buttons/button';
 
 export default function Nav() {
+    const { isLoggedIn } = useUser();
+
     return (
         <>
             <div className={styles['link-container']}>
@@ -10,7 +14,7 @@ export default function Nav() {
                 <PageLink href={'/#description'} label={'O stronie'} />
             </div>
             <div className={styles['user-container']}>
-                <User />
+                {isLoggedIn() ? <AccountDropdown /> : <Button label="Zaloguj się" icon="log in" link="/users/login" />}
             </div>
         </>
     );