+++ src/hooks/use-header-links.ts
@@ -8,7 +8,7 @@ import SearchIcon from '@mui/icons-material/Search';
 import VpnKeyIcon from '@mui/icons-material/VpnKey';
 import ListIcon from '@mui/icons-material/List';
 import AddIcon from '@mui/icons-material/Add';
-import WorkIcon from '@mui/icons-material/Work';
+import ManageAccountsIcon from '@mui/icons-material/ManageAccounts';
 import { useUser } from '@/src/hooks/use-user';
 import { useMemo } from 'react';
 
@@ -16,6 +16,7 @@ interface LinkData {
     link: string;
     label: string;
     Icon: OverridableComponent<SvgIconTypeMap> & { muiName: string };
+    includesDropdown?: boolean;
 }
 
 export function useHeaderLinks() {
@@ -27,13 +28,13 @@ export function useHeaderLinks() {
         ...(isLoggedIn()
             ? [
                 { link: '/recommendations', label: 'Propozycje dań', Icon: ListIcon },
-                { link: '/dishes/create', label: 'Stwórz nowe danie', Icon: AddIcon },
-                { link: '/users/change-password', label: 'Zmiana hasła', Icon: VpnKeyIcon }
+                { link: '/dishes/create', label: 'Stwórz nowe danie', Icon: AddIcon, includesDropdown: true },
+                { link: '/users/change-password', label: 'Zmiana hasła', Icon: VpnKeyIcon, includesDropdown: true }
             ]
             : []
         ),
         ...(hasCapability()
-            ? [{ link: '/manage', label: 'Zarządzaj', Icon: WorkIcon }]
+            ? [{ link: '/manage', label: 'Zarządzaj', Icon: ManageAccountsIcon, includesDropdown: true }]
             : []
         ),
         { link: '/#description', label: 'O stronie', Icon: InfoIcon }